Fix optimistic updates on proposal cards
